requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==21.2.0
billiard==3.6.3.0
cairocffi==1.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
Django==3.0.5
djangorestframework==3.11.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
kombu==4.6.8
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.9.5
pytest==6.2.4
pytz==2019.3
requests==2.25.1
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.2
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0
